payment 서비스 구성 정리
