--- athena/src/main/java/goorm/athena/domain/payment/controller/PaymentControllerImpl3.java
@@ -1,4 +1,0 @@
-package goorm.athena.domain.payment.controller;
-
-public class PaymentControllerImpl3 {
-}
