--- athena/src/main/java/goorm/athena/domain/payment/repository/PaymentRepository1.java
@@ -1,11 +1,0 @@
-package goorm.athena.domain.payment.repository;
-
-import goorm.athena.domain.payment.entity.Payment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PaymentRepository1 extends JpaRepository<Payment, Long> {
-    Optional<Payment> findByOrderId(Long orderId);
-
-}
