--- athena/src/main/java/goorm/athena/domain/payment/service/V3/KakaoService3.java
@@ -1,4 +1,0 @@
-package goorm.athena.domain.payment.service.V3;
-
-public class KakaoService3 {
-}
