+++ athena/src/main/java/goorm/athena/domain/payment/service/V4/RedisStockDeductionResult.java
@@ -1,0 +1,16 @@
+//package goorm.athena.domain.payment.service.V4;
+//
+//import lombok.Getter;
+//
+//import java.util.Map;
+//@Getter
+//public class RedisStockDeductionResult {
+//
+//    private final boolean success;
+//    private final Map<String, Integer> deductedStocks;
+//
+//    public RedisStockDeductionResult(boolean success, Map<String, Integer> deductedStocks) {
+//        this.success = success;
+//        this.deductedStocks = deductedStocks;
+//    }
+//}
